requirements.txt
pandas
numpy
scikit-learn
streamlit
kafka-python
psycopg2-binary
pyarrow
joblib

